Get Services from server
