+++ mockEnd/db.json
@@ -1,11 +1,25 @@
 {
   "Service": [
     {
-      "name": "string",
+      "name": "Box",
       "description": "string",
       "price": 1000
+    },
+    {
+      "name": "Sex",
+      "description": "string",
+      "price": 2000
+    },
+    {
+      "name": "Flex",
+      "description": "string",
+      "price": 3000
+    },
+    {
+      "name": "Kung-fu",
+      "description": "string",
+      "price": 4000
     }
-
   ],
   "Coach": [
     {
