--- src/Booking/Booking.js
@@ -1,10 +1,0 @@
-import React from "react";
-import styles from "./booking.module.css";
-
-class Booking extends React.Component {
-  render() {
-    return <h1>Booking</h1>;
-  }
-}
-
-export default Booking;
