--- src/Coaches/Coaches.js
@@ -1,10 +1,0 @@
-import React from "react";
-import styles from "./coaches.module.css";
-
-class Coaches extends React.Component {
-  render() {
-    return <h1>Coaches</h1>;
-  }
-}
-
-export default Coaches;
