--- src/Input/input.css
@@ -1,129 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Roboto&display=swap");
-
-* {
-  font-family: "Roboto";
-}
-
-:root {
-  --omrs-color-ink-lowest-contrast: rgba(255, 255, 255, 0.18);
-  --omrs-color-ink-low-contrast: rgba(60, 60, 67, 0.3);
-  --omrs-color-ink-medium-contrast: rgb(255, 255, 255);
-  --omrs-color-interaction: #f5ff01;
-  --omrs-color-interaction-minus-two: #6666661e;
-  --omrs-color-danger: #b50706;
-  --omrs-color-bg-low-contrast: #eff1f2;
-  --omrs-color-ink-high-contrast: #ffffff;
-  --omrs-color-bg-high-contrast: rgba(0, 0, 0, 0);
-}
-
-input:-webkit-autofill,
-input:-webkit-autofill:hover,
-input:-webkit-autofill:focus,
-input:-webkit-autofill:active {
-  caret-color: #fff; /* Pour le I quand on édite */
-  color: #fff;
-  background: #161616;
-  /* webkit autofill */
-  -webkit-text-fill-color: #fff; /* Surcharge la font color d'autofill */
-  -webkit-background-clip: text; /* Supprime le background autofill, utile pour le border radius */
-  box-shadow: 0 0 0 60px #161616 inset; /* Ajoute un fake background à base d'ombrage aplatit */
-}
-
-div.omrs-input-group {
-  margin-bottom: 1.5rem;
-  position: relative;
-  width: 20.4375rem;
-}
-
-.omrs-input-underlined > input {
-  border: none;
-  background: none;
-  border-bottom: 0.125rem solid var(--omrs-color-ink-medium-contrast);
-  width: 90%;
-  height: 2.5rem;
-  font-size: 1.0625rem;
-  padding-left: 0.875rem;
-  line-height: 147.6%;
-  padding-top: 0.825rem;
-  padding-bottom: 0.5rem;
-  color: rgb(255, 255, 255);
-}
-
-.omrs-input-underlined > input:focus {
-  outline: none;
-  background: #2424247c;
-  color: #f5ff01;
-}
-
-.omrs-input-underlined > .omrs-input-label {
-  position: absolute;
-  top: 0.9375rem;
-  left: 0.875rem;
-  line-height: 147.6%;
-  color: var(--omrs-color-ink-medium-contrast);
-  transition: top 0.2s;
-}
-
-.omrs-input-underlined > input:hover {
-  background: var(--omrs-color-interaction-minus-two);
-}
-
-.omrs-input-underlined > input:focus + .omrs-input-label,
-.omrs-input-underlined > input:valid + .omrs-input-label {
-  top: 0;
-  margin-top: 0;
-  font-size: 0.9375rem;
-  margin-bottom: 32px;
-  background: none;
-}
-
-.omrs-input-underlined:not(.omrs-input-danger)
-  > input:focus
-  + .omrs-input-label {
-  color: var(--omrs-color-interaction);
-}
-
-.omrs-input-underlined:not(.omrs-input-danger) > input:focus {
-  border-color: var(--omrs-color-interaction);
-}
-
-.omrs-input-underlined:not(.omrs-input-danger) > input:focus ~ svg {
-  fill: var(--omrs-color-ink-high-contrast);
-}
-
-/** DISABLED **/
-
-.omrs-input-underlined > input:disabled {
-  cursor: not-allowed;
-}
-
-.omrs-input-underlined > input:disabled + .omrs-input-label,
-.omrs-input-underlined > input:disabled ~ .omrs-input-helper {
-  color: var(--omrs-color-ink-low-contrast);
-}
-
-.omrs-input-underlined > input:disabled ~ svg {
-  fill: var(--omrs-color-ink-low-contrast);
-}
-
-/** DANGER **/
-
-.omrs-input-underlined.omrs-input-danger > .omrs-input-label,
-.omrs-input-underlined.omrs-input-danger > .omrs-input-helper {
-  color: var(--omrs-color-danger);
-}
-
-.omrs-input-danger > svg {
-  fill: var(--omrs-color-danger);
-}
-
-.omrs-input-danger > input {
-  border-color: var(--omrs-color-danger);
-}
-
-.omrs-input-underlined > input {
-  background: var(--omrs-color-bg-high-contrast);
-}
-.omrs-input-filled > input {
-  background: var(--omrs-color-bg-low-contrast);
-}
