--- src/Services/services.module.css
@@ -1,124 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: row;
-  gap: 60px;
-
-  width: 70vw;
-  margin: 0 auto;
-}
-
-.addBtn {
-  outline: none;
-  border: none;
-  background: none;
-
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 15px;
-
-  padding: 0.4em 4em;
-  margin: 15px 0 5px 0;
-  width: inherit;
-  height: fit-content;
-
-  border-radius: 15px;
-  box-sizing: border-box;
-  background: #86868621;
-
-  text-decoration: none;
-  font-family: "Roboto", sans-serif;
-  font-weight: 300;
-  font-size: 1.3rem;
-  color: #ffffff;
-  text-align: center;
-  transition: all 0.2s;
-}
-
-.addBtn:hover {
-  color: #000000;
-  background-color: #f5ff01;
-  border-color: #f5ff01;
-
-  cursor: pointer;
-}
-
-.addBtnText {
-  margin: auto;
-}
-
-.addBtnPlus {
-  height: 25px;
-  width: 25px;
-
-  margin: auto;
-  padding: auto;
-  border: 1px solid;
-  border-radius: 50%;
-
-  font-size: 20px;
-  font-style: normal;
-}
-
-/*//////////////////////////////////*/
-
-.leftContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-
-  width: 250px;
-}
-
-.rightContainer {
-  margin-top: 100px;
-}
-
-.resultList {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-
-  border: 1px solid #f5ff01;
-  border-radius: 20px;
-  background: #01010152;
-
-  width: inherit;
-  height: 70vh;
-  padding-top: 15px;
-  padding-bottom: 15px;
-
-  -ms-overflow-style: none;
-  overflow-y: scroll;
-  scrollbar-width: none;
-  scroll-behavior: smooth;
-}
-
-.resultList::-webkit-scrollbar {
-  display: none;
-}
-
-.resultList > p {
-  width: 60%;
-
-  font-size: 1.4rem;
-  text-align: center;
-  color: #f5ff01;
-  text-transform: uppercase;
-
-  padding-bottom: 4px;
-  cursor: pointer;
-}
-
-.resultList > p::after {
-  display: block;
-  content: "";
-  border-bottom: solid 1px #f5ff01;
-  transform: scaleX(0);
-  transition: transform 250ms ease-in-out;
-}
-
-.resultList > p:hover:after {
-  transform: scaleX(1);
-}
