--- src/Slots/Slots.js
@@ -1,10 +1,0 @@
-import React from "react";
-import styles from "./slots.module.css";
-
-class Slots extends React.Component {
-  render() {
-    return <h1>Slots</h1>;
-  }
-}
-
-export default Slots;
