+++ src/components/Clients/Clients.js
@@ -1,0 +1,10 @@
+import React from "react";
+import styles from "./clients.module.css";
+
+class Clients extends React.Component {
+  render() {
+    return <h1>Clients</h1>;
+  }
+}
+
+export default Clients;
