+++ src/index.js
@@ -1,12 +1,28 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import {render} from 'react-dom';
+import {applyMiddleware, compose, createStore} from "redux";
+import {Provider} from "react-redux";
+import thunk from "redux-thunk";
+
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import {reducer} from "./redux/reducer";
 
-ReactDOM.render(
+const store = createStore(
+    reducer,
+    compose(
+        applyMiddleware(
+            thunk
+        ),
+        window?.__REDUX_DEVTOOLS_EXTENSION__()
+    )
+)
+render(
   <React.StrictMode>
-    <App />
+      <Provider store={store}>
+          <App />
+      </Provider>
   </React.StrictMode>,
   document.getElementById('root')
 );
