+++ src/redux/axios.js
@@ -1,0 +1,5 @@
+import axios from "axios";
+
+export default axios.create({
+    baseURL: 'http://localhost:4096',
+})
