+++ src/redux/reducer.js
@@ -1,0 +1,6 @@
+import {combineReducers} from "redux";
+import {serviceReducer} from "./serviceReducer";
+
+export const reducer = combineReducers({
+    service: serviceReducer
+})
