+++ src/redux/types.js
@@ -1,0 +1,3 @@
+export const GET_SERVICES  = 'SERVICES/GET_SERVICES'
+export const ADD_SERVICE  = 'SERVICES/ADD_SERVICE'
+export const RESET_SERVICES  = 'SERVICES/RESET_SERVICES'
